requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
tzdata==2025.1
psycopg2-binary>=2.8
redis>=3.5
celery>=5.0
djangorestframework==3.14.0
django-allauth==0.54.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
python-dotenv==1.0.0
django-anymail==12.0
pillow==11.1.0
django-filter==25.1